Added Spinboxes and if-else statment
I replaced the Entry widget with a Spinbox.
I tested our code and introduced an if-else statement to display a warning pop-up window if the user enters an letter. However, we still need to add and elif statement to display a separate warning message if the user enters a symbol or a space, and math domain errors need fixing too (due to the domain of arcsine).